added

+++ client/src/pages/Profile.jsx
@@ -17,7 +17,7 @@ import {
   signOutUserStart,
 } from "../redux/user/userSlice";
 import { useDispatch } from "react-redux";
-import { Link } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
 export default function Profile() {
   const fileRef = useRef(null);
   const { currentUser, loading, error } = useSelector((state) => state.user);
@@ -29,7 +29,7 @@ export default function Profile() {
   const [showListingsError, setShowListingsError] = useState(false);
   const [userListings, setUserListings] = useState([]);
   const dispatch = useDispatch();
-
+  const navigate = useNavigate();
   // firebase storage
   // allow read;
   // allow write: if
@@ -84,6 +84,10 @@ export default function Profile() {
         body: JSON.stringify(formData),
       });
       const data = await res.json();
+      if(data==='Unauthorised'){
+        navigate('/errorhandle')
+        return;
+      }
       if (data.success === false) {
         dispatch(updateUserFailure(data.message));
         return;
@@ -102,6 +106,10 @@ export default function Profile() {
         method: "DELETE",
       });
       const data = await res.json();
+      if(data==='Unauthorised'){
+        navigate('/errorhandle')
+        return;
+      }
       if (data.success === false) {
         dispatch(deleteUserFailure(data.message));
         return;
@@ -117,6 +125,10 @@ export default function Profile() {
       dispatch(signOutUserStart());
       const res = await fetch("/api/auth/signout");
       const data = await res.json();
+      if(data==='Unauthorised'){
+        navigate('/errorhandle')
+        return;
+      }
       if (data.success === false) {
         dispatch(deleteUserFailure(data.message));
         return;
@@ -132,6 +144,10 @@ export default function Profile() {
       setShowListingsError(false);
       const res = await fetch(`/api/user/listings/${currentUser._id}`);
       const data = await res.json();
+      if(data==='Unauthorised'){
+        navigate('/errorhandle')
+        return;
+      }
       if (data.success === false) {
         setShowListingsError(true);
         return;
@@ -150,6 +166,10 @@ export default function Profile() {
         method: 'DELETE',
       });
       const data = await res.json();
+      if(data==='Unauthorised'){
+        navigate('/errorhandle')
+        return;
+      }
       if (data.success === false) {
         console.log(data.message);
         return;
